requirements.txt
attrs==25.3.0
beautifulsoup4==4.13.4
bs4==0.0.2
certifi==2025.4.26
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.2
cssselect==1.3.0
h11==0.16.0
idna==3.10
lxml==6.0.0
lxml_html_clean==0.4.2
numpy==2.2.6
outcome==1.3.0.post0
packaging==25.0
pandas==2.2.3
pdf2image==1.17.0
pillow==11.3.0
pycparser==2.22
PySocks==1.7.1
pytesseract==0.3.13
python-dateutil==2.9.0.post0
pytz==2025.2
readability-lxml==0.8.4.1
requests==2.32.4
selenium==4.32.0
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.7
trio==0.30.0
trio-websocket==0.12.2
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
websocket-client==1.8.0
wsproto==1.2.0
